requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
faker==20.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
httpx==0.25.2
aiofiles==23.2.1
asyncpg==0.29.0
python-dotenv==1.0.0
orjson==3.9.10
twilio==8.11.0